+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - feature: improve git debugging
 - feature: add troubleshooting menu item
 - cleanup: refactored cli execute
+- cleanup: remove death quiqr-cloud code
 
 ## 0.16.0 (2023-04-24)
 - feature: add back button in prefs toolbar